#28 create new TestITEndToEnd and move the dollar token full cycle test into there
